Attempting to break apart and encapsulate logic
* We have many different types of run modes and behaviour is subtly
  dependent on each other.
* In this approach, everything will be tied to gether that needs to be
  and all the subtly will be explict, such as resolving sources must
come after decompiling a binary.
* When decompiling we have to take into account each type of archive and
  where they should be putting files
* When resolving sources, we need specific build tool version.
* Right now, during init, we do all of this work with deps, and then
  IIUC re-do it in the dependency code, We shouldn't need to do this I
think.

* I also want to change the decompile behaviour because fernflower can
  decompile the whole archive, we don't have to go file by file and
decompile each class.
