+++ external-providers/java-external-provider/pkg/java_external_provider/war.go
@@ -1,0 +1,10 @@
+package java
+
+type war struct {
+	// Fields for War file
+
+}
+
+func NewWar() (Source, error) {
+	return &war{}
+}
